refactor metadata keyword workflows by file
