+++ internal/cli/metadata/keywords.go
@@ -2,17 +2,13 @@ package metadata
 
 import (
 	"context"
-	"encoding/csv"
-	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 	"sort"
 	"strings"
-	"time"
 
 	"github.com/peterbourgon/ff/v3/ffcli"
 
@@ -20,25 +16,6 @@ import (
 	"github.com/rudrankriyam/App-Store-Connect-CLI/internal/cli/shared"
 	"github.com/rudrankriyam/App-Store-Connect-CLI/internal/validation"
 )
-
-const (
-	keywordImportFormatAuto     = "auto"
-	keywordImportFormatCSV      = "csv"
-	keywordImportFormatJSON     = "json"
-	keywordImportFormatText     = "text"
-	keywordImportFormatAstroCSV = "astro-csv"
-)
-
-var keywordPlanFields = []string{"keywords"}
-
-type metadataKeywordImportParser func([]byte, string) (metadataKeywordImportedData, error)
-
-var metadataKeywordImportFormats = map[string]metadataKeywordImportParser{
-	keywordImportFormatCSV:      parseMetadataKeywordCSV,
-	keywordImportFormatJSON:     parseMetadataKeywordJSON,
-	keywordImportFormatText:     parseMetadataKeywordText,
-	keywordImportFormatAstroCSV: parseMetadataKeywordAstroCSV,
-}
 
 // MetadataKeywordFileResult describes one local keyword file change.
 type MetadataKeywordFileResult struct {
@@ -63,110 +40,6 @@ type MetadataKeywordIssue struct {
 	Limit        int    `json:"limit,omitempty"`
 }
 
-// MetadataKeywordsImportResult describes one import run.
-type MetadataKeywordsImportResult struct {
-	Dir                 string                      `json:"dir"`
-	Version             string                      `json:"version"`
-	Input               string                      `json:"input"`
-	Format              string                      `json:"format"`
-	DryRun              bool                        `json:"dryRun"`
-	Valid               bool                        `json:"valid"`
-	DetectedLocales     []string                    `json:"detectedLocales"`
-	Results             []MetadataKeywordFileResult `json:"results"`
-	Issues              []MetadataKeywordIssue      `json:"issues,omitempty"`
-	SideDataRecordCount int                         `json:"sideDataRecordCount,omitempty"`
-	SideDataReportPath  string                      `json:"sideDataReportPath,omitempty"`
-}
-
-// MetadataKeywordsLocalizeResult describes one localization-copy run.
-type MetadataKeywordsLocalizeResult struct {
-	Dir                 string                      `json:"dir"`
-	Version             string                      `json:"version"`
-	FromLocale          string                      `json:"fromLocale"`
-	DryRun              bool                        `json:"dryRun"`
-	Valid               bool                        `json:"valid"`
-	DetectedLocales     []string                    `json:"detectedLocales"`
-	Results             []MetadataKeywordFileResult `json:"results"`
-	Issues              []MetadataKeywordIssue      `json:"issues,omitempty"`
-	SideDataRecordCount int                         `json:"sideDataRecordCount,omitempty"`
-	SideDataReportPath  string                      `json:"sideDataReportPath,omitempty"`
-}
-
-// MetadataKeywordsWarning highlights submit-readiness risk during keyword creates.
-type MetadataKeywordsWarning struct {
-	Action        string   `json:"action"`
-	Locale        string   `json:"locale"`
-	Message       string   `json:"message"`
-	MissingFields []string `json:"missingFields,omitempty"`
-}
-
-// MetadataKeywordsMutationResult describes one remote keyword mutation attempt.
-type MetadataKeywordsMutationResult struct {
-	Locale         string `json:"locale"`
-	Action         string `json:"action"`
-	Status         string `json:"status"`
-	LocalizationID string `json:"localizationId,omitempty"`
-	Error          string `json:"error,omitempty"`
-}
-
-// MetadataKeywordsPlanResult describes keyword-only remote changes.
-type MetadataKeywordsPlanResult struct {
-	AppID               string                           `json:"appId"`
-	Version             string                           `json:"version"`
-	VersionID           string                           `json:"versionId"`
-	Dir                 string                           `json:"dir"`
-	DryRun              bool                             `json:"dryRun"`
-	Applied             bool                             `json:"applied,omitempty"`
-	Total               int                              `json:"total,omitempty"`
-	Succeeded           int                              `json:"succeeded,omitempty"`
-	Failed              int                              `json:"failed,omitempty"`
-	FailureArtifactPath string                           `json:"failureArtifactPath,omitempty"`
-	Adds                []PlanItem                       `json:"adds"`
-	Updates             []PlanItem                       `json:"updates"`
-	APICalls            []PlanAPICall                    `json:"apiCalls,omitempty"`
-	Actions             []ApplyAction                    `json:"actions,omitempty"`
-	Results             []MetadataKeywordsMutationResult `json:"results,omitempty"`
-	Warnings            []MetadataKeywordsWarning        `json:"warnings,omitempty"`
-}
-
-// MetadataKeywordsSyncResult combines import and remote planning/apply.
-type MetadataKeywordsSyncResult struct {
-	Import MetadataKeywordsImportResult `json:"import"`
-	Plan   *MetadataKeywordsPlanResult  `json:"plan,omitempty"`
-}
-
-type metadataKeywordsImportOptions struct {
-	Dir                string
-	Version            string
-	Input              string
-	Format             string
-	DefaultLocale      string
-	DryRun             bool
-	Overwrite          bool
-	SideDataReportFile string
-}
-
-type metadataKeywordsLocalizeOptions struct {
-	Dir           string
-	Version       string
-	FromLocale    string
-	TargetLocales []string
-	DryRun        bool
-	Overwrite     bool
-}
-
-type metadataKeywordsPlanOptions struct {
-	AppID                string
-	Version              string
-	Platform             string
-	Dir                  string
-	DryRun               bool
-	Apply                bool
-	Confirm              bool
-	FailureArtifactScope string
-	LocalState           map[string]keywordLocalState
-}
-
 type keywordLocalState struct {
 	locale string
 	file   string
@@ -174,63 +47,10 @@ type keywordLocalState struct {
 	patch  versionLocalPatch
 }
 
-type keywordImportPayload struct {
-	states map[string]keywordLocalState
-	result MetadataKeywordsImportResult
-}
-
 type metadataKeywordFieldDetails struct {
 	field      string
 	count      int
 	duplicates []string
-}
-
-type metadataKeywordImportedData struct {
-	locales  map[string][]string
-	sideData []MetadataKeywordSideDataRecord
-}
-
-type metadataKeywordsApplyFailureArtifact struct {
-	AppID       string                           `json:"appId"`
-	Version     string                           `json:"version"`
-	VersionID   string                           `json:"versionId"`
-	Dir         string                           `json:"dir"`
-	Total       int                              `json:"total"`
-	Succeeded   int                              `json:"succeeded"`
-	Failed      int                              `json:"failed"`
-	GeneratedAt string                           `json:"generatedAt"`
-	Results     []MetadataKeywordsMutationResult `json:"results"`
-}
-
-type metadataKeywordsApplySummary struct {
-	Total     int
-	Succeeded int
-	Failed    int
-	Actions   []ApplyAction
-	Results   []MetadataKeywordsMutationResult
-}
-
-func shouldPrintMetadataKeywordsPlanResult(result MetadataKeywordsPlanResult, err error) bool {
-	if err == nil {
-		return true
-	}
-	return result.Total > 0 || result.Succeeded > 0 || result.Failed > 0 || len(result.Results) > 0 || result.FailureArtifactPath != ""
-}
-
-// MetadataKeywordSideDataRecord captures non-publishable research fields from imports.
-type MetadataKeywordSideDataRecord struct {
-	Locale   string         `json:"locale,omitempty"`
-	Keywords []string       `json:"keywords,omitempty"`
-	Fields   map[string]any `json:"fields"`
-}
-
-// MetadataKeywordSideDataArtifact is the persisted side-data report format.
-type MetadataKeywordSideDataArtifact struct {
-	Dir     string                          `json:"dir"`
-	Version string                          `json:"version"`
-	Input   string                          `json:"input"`
-	Format  string                          `json:"format"`
-	Records []MetadataKeywordSideDataRecord `json:"records"`
 }
 
 // MetadataKeywordsCommand returns the canonical metadata keywords subtree.
@@ -280,892 +100,6 @@ Examples:
 	}
 }
 
-// MetadataKeywordsImportCommand returns the keywords import subcommand.
-func MetadataKeywordsImportCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("metadata keywords import", flag.ExitOnError)
-
-	dir := fs.String("dir", "", "Metadata root directory (required)")
-	version := fs.String("version", "", "App version string (for example 1.2.3)")
-	input := fs.String("input", "", "Import file path or - for stdin (required)")
-	format := fs.String("format", keywordImportFormatAuto, "Input format: auto, csv, json, text, or astro-csv")
-	locale := fs.String("locale", "", "Default locale for inputs without a locale column/field")
-	sideDataReportFile := fs.String("side-data-report-file", "", "Optional path to write side-data report JSON when research fields are present")
-	dryRun := fs.Bool("dry-run", false, "Preview local file changes without writing files")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "import",
-		ShortUsage: "asc metadata keywords import --dir \"./metadata\" --version \"1.2.3\" --input \"./keywords.csv\" [flags]",
-		ShortHelp:  "Import provider keyword exports into canonical metadata files.",
-		LongHelp: `Import provider keyword exports into canonical metadata files.
-
-Supported input formats:
-  - csv: header-based rows with locale + keywords/keyword/term columns
-  - json: locale-keyed maps, arrays of localization objects, or a single localization object
-  - text: a plain comma/newline-separated keyword list (requires --locale)
-  - astro-csv: Astro keyword export CSV using the documented Keyword column (requires --locale unless locale data is present)
-
-Examples:
-  asc metadata keywords import --dir "./metadata" --version "1.2.3" --locale "en-US" --input "./keywords.csv"
-  asc metadata keywords import --dir "./metadata" --version "1.2.3" --format json --input "./keywords.json"
-  asc metadata keywords import --dir "./metadata" --version "1.2.3" --format text --locale "fr-FR" --input "./keywords.txt" --dry-run`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			if len(args) > 0 {
-				return shared.UsageError("metadata keywords import does not accept positional arguments")
-			}
-			result, err := executeMetadataKeywordsImport(metadataKeywordsImportOptions{
-				Dir:                *dir,
-				Version:            *version,
-				Input:              *input,
-				Format:             *format,
-				DefaultLocale:      *locale,
-				DryRun:             *dryRun,
-				Overwrite:          true,
-				SideDataReportFile: *sideDataReportFile,
-			})
-			if err != nil {
-				if errors.Is(err, flag.ErrHelp) {
-					return err
-				}
-				return fmt.Errorf("metadata keywords import: %w", err)
-			}
-			if err := shared.PrintOutputWithRenderers(
-				result,
-				*output.Output,
-				*output.Pretty,
-				func() error {
-					return printMetadataKeywordFileResultTable("Keyword Import", result.Results, result.DetectedLocales, result.Issues, result.Dir, result.Version, result.DryRun, result.SideDataRecordCount, result.SideDataReportPath)
-				},
-				func() error {
-					return printMetadataKeywordFileResultMarkdown("Keyword Import", result.Results, result.DetectedLocales, result.Issues, result.Dir, result.Version, result.DryRun, result.SideDataRecordCount, result.SideDataReportPath)
-				},
-			); err != nil {
-				return err
-			}
-			if !result.Valid {
-				return shared.NewReportedError(fmt.Errorf("metadata keywords import: found %d issue(s)", len(result.Issues)))
-			}
-			return nil
-		},
-	}
-}
-
-// MetadataKeywordsPlanCommand returns the keywords plan subcommand.
-func MetadataKeywordsPlanCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("metadata keywords plan", flag.ExitOnError)
-
-	appID := fs.String("app", "", "App Store Connect app ID (or ASC_APP_ID env)")
-	version := fs.String("version", "", "App version string (for example 1.2.3)")
-	platform := fs.String("platform", "", "Optional platform: IOS, MAC_OS, TV_OS, or VISION_OS")
-	dir := fs.String("dir", "", "Metadata root directory (required)")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "plan",
-		ShortUsage: "asc metadata keywords plan --app \"APP_ID\" --version \"1.2.3\" --dir \"./metadata\" [flags]",
-		ShortHelp:  "Preview keyword-only changes against App Store Connect.",
-		LongHelp: `Preview keyword-only changes against App Store Connect.
-
-This command reads local canonical metadata files, looks only at the version
-localization ` + "`keywords`" + ` field, and builds a non-mutating plan.
-
-Examples:
-  asc metadata keywords plan --app "APP_ID" --version "1.2.3" --dir "./metadata"
-  asc metadata keywords plan --app "APP_ID" --version "1.2.3" --platform IOS --dir "./metadata"`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			return runMetadataKeywordsPlanLikeCommand(ctx, args, "metadata keywords plan", metadataKeywordsPlanOptions{
-				AppID:    *appID,
-				Version:  *version,
-				Platform: *platform,
-				Dir:      *dir,
-				DryRun:   true,
-			}, output)
-		},
-	}
-}
-
-// MetadataKeywordsDiffCommand returns the keywords diff subcommand.
-func MetadataKeywordsDiffCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("metadata keywords diff", flag.ExitOnError)
-
-	appID := fs.String("app", "", "App Store Connect app ID (or ASC_APP_ID env)")
-	version := fs.String("version", "", "App version string (for example 1.2.3)")
-	platform := fs.String("platform", "", "Optional platform: IOS, MAC_OS, TV_OS, or VISION_OS")
-	dir := fs.String("dir", "", "Metadata root directory (required)")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "diff",
-		ShortUsage: "asc metadata keywords diff --app \"APP_ID\" --version \"1.2.3\" --dir \"./metadata\" [flags]",
-		ShortHelp:  "Diff local canonical keywords against App Store Connect.",
-		LongHelp: `Diff local canonical keywords against App Store Connect.
-
-This is a keyword-focused alias of the planning flow, intended for human review
-of local-vs-remote keyword changes before apply.
-
-Examples:
-  asc metadata keywords diff --app "APP_ID" --version "1.2.3" --dir "./metadata"
-  asc metadata keywords diff --app "APP_ID" --version "1.2.3" --platform IOS --dir "./metadata"`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			return runMetadataKeywordsPlanLikeCommand(ctx, args, "metadata keywords diff", metadataKeywordsPlanOptions{
-				AppID:    *appID,
-				Version:  *version,
-				Platform: *platform,
-				Dir:      *dir,
-				DryRun:   true,
-			}, output)
-		},
-	}
-}
-
-func runMetadataKeywordsPlanLikeCommand(
-	ctx context.Context,
-	args []string,
-	errorPrefix string,
-	opts metadataKeywordsPlanOptions,
-	output shared.OutputFlags,
-) error {
-	if len(args) > 0 {
-		return shared.UsageError(fmt.Sprintf("%s does not accept positional arguments", errorPrefix))
-	}
-	result, err := executeMetadataKeywordsPlan(ctx, opts)
-	if err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return err
-		}
-		return fmt.Errorf("%s: %w", errorPrefix, err)
-	}
-	return shared.PrintOutputWithRenderers(
-		result,
-		*output.Output,
-		*output.Pretty,
-		func() error { return printMetadataKeywordsPlanTable(result) },
-		func() error { return printMetadataKeywordsPlanMarkdown(result) },
-	)
-}
-
-// MetadataKeywordsLocalizeCommand returns the keywords localize subcommand.
-func MetadataKeywordsLocalizeCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("metadata keywords localize", flag.ExitOnError)
-
-	dir := fs.String("dir", "", "Metadata root directory (required)")
-	version := fs.String("version", "", "App version string (for example 1.2.3)")
-	fromLocale := fs.String("from-locale", "", "Source locale to copy keywords from (required)")
-	toLocales := fs.String("to-locales", "", "Target locales (comma-separated, required)")
-	overwrite := fs.Bool("overwrite", false, "Overwrite existing target keyword fields")
-	dryRun := fs.Bool("dry-run", false, "Preview local file changes without writing files")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "localize",
-		ShortUsage: "asc metadata keywords localize --dir \"./metadata\" --version \"1.2.3\" --from-locale \"en-US\" --to-locales \"fr-FR,de-DE\" [flags]",
-		ShortHelp:  "Copy one locale's canonical keywords into other locales.",
-		LongHelp: `Copy one locale's canonical keywords into other locales.
-
-This command copies the canonical keyword field from one locale into one or
-more target locale files. It does not translate terms; it seeds target locale
-files so they can be reviewed and refined before apply.
-
-Examples:
-  asc metadata keywords localize --dir "./metadata" --version "1.2.3" --from-locale "en-US" --to-locales "fr-FR,de-DE"
-  asc metadata keywords localize --dir "./metadata" --version "1.2.3" --from-locale "en-US" --to-locales "it,es-MX" --overwrite --dry-run`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			if len(args) > 0 {
-				return shared.UsageError("metadata keywords localize does not accept positional arguments")
-			}
-			targets := shared.SplitUniqueCSV(*toLocales)
-			result, err := executeMetadataKeywordsLocalize(metadataKeywordsLocalizeOptions{
-				Dir:           *dir,
-				Version:       *version,
-				FromLocale:    *fromLocale,
-				TargetLocales: targets,
-				DryRun:        *dryRun,
-				Overwrite:     *overwrite,
-			})
-			if err != nil {
-				if errors.Is(err, flag.ErrHelp) {
-					return err
-				}
-				return fmt.Errorf("metadata keywords localize: %w", err)
-			}
-			if err := shared.PrintOutputWithRenderers(
-				result,
-				*output.Output,
-				*output.Pretty,
-				func() error {
-					return printMetadataKeywordFileResultTable("Keyword Localize", result.Results, result.DetectedLocales, result.Issues, result.Dir, result.Version, result.DryRun, 0, "")
-				},
-				func() error {
-					return printMetadataKeywordFileResultMarkdown("Keyword Localize", result.Results, result.DetectedLocales, result.Issues, result.Dir, result.Version, result.DryRun, 0, "")
-				},
-			); err != nil {
-				return err
-			}
-			if !result.Valid {
-				return shared.NewReportedError(fmt.Errorf("metadata keywords localize: found %d issue(s)", len(result.Issues)))
-			}
-			return nil
-		},
-	}
-}
-
-// MetadataKeywordsApplyCommand returns the keywords apply subcommand.
-func MetadataKeywordsApplyCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("metadata keywords apply", flag.ExitOnError)
-
-	appID := fs.String("app", "", "App Store Connect app ID (or ASC_APP_ID env)")
-	version := fs.String("version", "", "App version string (for example 1.2.3)")
-	platform := fs.String("platform", "", "Optional platform: IOS, MAC_OS, TV_OS, or VISION_OS")
-	dir := fs.String("dir", "", "Metadata root directory (required)")
-	confirm := fs.Bool("confirm", false, "Confirm remote keyword mutations")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "apply",
-		ShortUsage: "asc metadata keywords apply --app \"APP_ID\" --version \"1.2.3\" --dir \"./metadata\" --confirm [flags]",
-		ShortHelp:  "Apply keyword-only metadata changes to App Store Connect.",
-		LongHelp: `Apply keyword-only metadata changes to App Store Connect.
-
-This command mutates only the version-localization ` + "`keywords`" + ` field.
-Other version metadata fields remain untouched by updates performed here.
-
-Examples:
-  asc metadata keywords apply --app "APP_ID" --version "1.2.3" --dir "./metadata" --confirm
-  asc metadata keywords apply --app "APP_ID" --version "1.2.3" --platform IOS --dir "./metadata" --confirm`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			if len(args) > 0 {
-				return shared.UsageError("metadata keywords apply does not accept positional arguments")
-			}
-			result, err := executeMetadataKeywordsPlan(ctx, metadataKeywordsPlanOptions{
-				AppID:                *appID,
-				Version:              *version,
-				Platform:             *platform,
-				Dir:                  *dir,
-				DryRun:               false,
-				Apply:                true,
-				Confirm:              *confirm,
-				FailureArtifactScope: "metadata-keywords-apply",
-			})
-			if err != nil && errors.Is(err, flag.ErrHelp) {
-				return err
-			}
-			if !shouldPrintMetadataKeywordsPlanResult(result, err) {
-				return fmt.Errorf("metadata keywords apply: %w", err)
-			}
-			if printErr := shared.PrintOutputWithRenderers(
-				result,
-				*output.Output,
-				*output.Pretty,
-				func() error { return printMetadataKeywordsPlanTable(result) },
-				func() error { return printMetadataKeywordsPlanMarkdown(result) },
-			); printErr != nil {
-				return printErr
-			}
-			if err != nil {
-				return fmt.Errorf("metadata keywords apply: %w", err)
-			}
-			if result.Failed > 0 {
-				return shared.NewReportedError(fmt.Errorf("metadata keywords apply: %d locale(s) failed", result.Failed))
-			}
-			return nil
-		},
-	}
-}
-
-// MetadataKeywordsSyncCommand returns the keywords sync subcommand.
-func MetadataKeywordsSyncCommand() *ffcli.Command {
-	fs := flag.NewFlagSet("metadata keywords sync", flag.ExitOnError)
-
-	appID := fs.String("app", "", "App Store Connect app ID (or ASC_APP_ID env)")
-	version := fs.String("version", "", "App version string (for example 1.2.3)")
-	platform := fs.String("platform", "", "Optional platform: IOS, MAC_OS, TV_OS, or VISION_OS")
-	dir := fs.String("dir", "", "Metadata root directory (required)")
-	input := fs.String("input", "", "Import file path or - for stdin (required)")
-	format := fs.String("format", keywordImportFormatAuto, "Input format: auto, csv, json, text, or astro-csv")
-	locale := fs.String("locale", "", "Default locale for inputs without a locale column/field")
-	sideDataReportFile := fs.String("side-data-report-file", "", "Optional path to write side-data report JSON when research fields are present")
-	dryRun := fs.Bool("dry-run", false, "Preview import and remote keyword changes without writing or mutating")
-	confirm := fs.Bool("confirm", false, "Confirm remote keyword mutations after import")
-	output := shared.BindOutputFlags(fs)
-
-	return &ffcli.Command{
-		Name:       "sync",
-		ShortUsage: "asc metadata keywords sync --app \"APP_ID\" --version \"1.2.3\" --dir \"./metadata\" --input \"./keywords.csv\" [flags]",
-		ShortHelp:  "Import keyword input and sync the resulting keyword plan.",
-		LongHelp: `Import keyword input and sync the resulting keyword plan.
-
-Workflow:
-  1. normalize provider input into canonical metadata keyword files
-  2. build a keyword-only remote plan for the imported locales
-  3. apply the remote changes only when --confirm is provided
-
-Without ` + "`--confirm`" + `, sync writes local files (unless ` + "`--dry-run`" + `)
-and returns a non-mutating remote plan.
-
-Examples:
-  asc metadata keywords sync --app "APP_ID" --version "1.2.3" --dir "./metadata" --input "./keywords.csv"
-  asc metadata keywords sync --app "APP_ID" --version "1.2.3" --platform IOS --dir "./metadata" --input "./keywords.json" --format json --confirm
-  asc metadata keywords sync --app "APP_ID" --version "1.2.3" --dir "./metadata" --format text --locale "en-US" --input "./keywords.txt" --dry-run`,
-		FlagSet:   fs,
-		UsageFunc: shared.DefaultUsageFunc,
-		Exec: func(ctx context.Context, args []string) error {
-			if len(args) > 0 {
-				return shared.UsageError("metadata keywords sync does not accept positional arguments")
-			}
-			resolvedAppID, versionValue, dirValue, platformValue, err := validateMetadataKeywordsRemoteInputs(*appID, *version, *dir, *platform)
-			if err != nil {
-				return fmt.Errorf("metadata keywords sync: %w", err)
-			}
-			importPayload, err := executeMetadataKeywordsImportWithState(metadataKeywordsImportOptions{
-				Dir:                *dir,
-				Version:            *version,
-				Input:              *input,
-				Format:             *format,
-				DefaultLocale:      *locale,
-				DryRun:             *dryRun,
-				Overwrite:          true,
-				SideDataReportFile: *sideDataReportFile,
-			})
-			if err != nil {
-				if errors.Is(err, flag.ErrHelp) {
-					return err
-				}
-				return fmt.Errorf("metadata keywords sync: %w", err)
-			}
-			if !importPayload.result.Valid {
-				result := MetadataKeywordsSyncResult{Import: importPayload.result}
-				if err := shared.PrintOutputWithRenderers(
-					result,
-					*output.Output,
-					*output.Pretty,
-					func() error { return printMetadataKeywordsSyncTable(result) },
-					func() error { return printMetadataKeywordsSyncMarkdown(result) },
-				); err != nil {
-					return err
-				}
-				return shared.NewReportedError(fmt.Errorf("metadata keywords sync: found %d import issue(s)", len(importPayload.result.Issues)))
-			}
-
-			planResult, err := executeMetadataKeywordsPlan(ctx, metadataKeywordsPlanOptions{
-				AppID:                resolvedAppID,
-				Version:              versionValue,
-				Platform:             platformValue,
-				Dir:                  dirValue,
-				DryRun:               *dryRun || !*confirm,
-				Apply:                !*dryRun && *confirm,
-				Confirm:              *confirm,
-				FailureArtifactScope: "metadata-keywords-sync",
-				LocalState:           importPayload.states,
-			})
-			if err != nil && errors.Is(err, flag.ErrHelp) {
-				return err
-			}
-			if !shouldPrintMetadataKeywordsPlanResult(planResult, err) {
-				return fmt.Errorf("metadata keywords sync: %w", err)
-			}
-
-			result := MetadataKeywordsSyncResult{
-				Import: importPayload.result,
-				Plan:   &planResult,
-			}
-			if printErr := shared.PrintOutputWithRenderers(
-				result,
-				*output.Output,
-				*output.Pretty,
-				func() error { return printMetadataKeywordsSyncTable(result) },
-				func() error { return printMetadataKeywordsSyncMarkdown(result) },
-			); printErr != nil {
-				return printErr
-			}
-			if err != nil {
-				return fmt.Errorf("metadata keywords sync: %w", err)
-			}
-			if planResult.Failed > 0 {
-				return shared.NewReportedError(fmt.Errorf("metadata keywords sync: %d locale(s) failed", planResult.Failed))
-			}
-			return nil
-		},
-	}
-}
-
-func validateMetadataKeywordsRemoteInputs(appID string, version string, dir string, platform string) (string, string, string, string, error) {
-	resolvedAppID := shared.ResolveAppID(appID)
-	if resolvedAppID == "" {
-		return "", "", "", "", shared.UsageError("--app is required (or set ASC_APP_ID)")
-	}
-
-	dirValue, versionValue, err := validateMetadataKeywordDirVersion(dir, version)
-	if err != nil {
-		return "", "", "", "", err
-	}
-
-	platformValue := strings.TrimSpace(platform)
-	if platformValue != "" {
-		normalizedPlatform, platformErr := shared.NormalizeAppStoreVersionPlatform(platformValue)
-		if platformErr != nil {
-			return "", "", "", "", shared.UsageError(platformErr.Error())
-		}
-		platformValue = normalizedPlatform
-	}
-
-	return resolvedAppID, versionValue, dirValue, platformValue, nil
-}
-
-func executeMetadataKeywordsImport(opts metadataKeywordsImportOptions) (MetadataKeywordsImportResult, error) {
-	payload, err := executeMetadataKeywordsImportWithState(opts)
-	if err != nil {
-		return MetadataKeywordsImportResult{}, err
-	}
-	return payload.result, nil
-}
-
-func executeMetadataKeywordsImportWithState(opts metadataKeywordsImportOptions) (keywordImportPayload, error) {
-	dirValue, versionValue, err := validateMetadataKeywordDirVersion(opts.Dir, opts.Version)
-	if err != nil {
-		return keywordImportPayload{}, err
-	}
-	inputValue := strings.TrimSpace(opts.Input)
-	if inputValue == "" {
-		return keywordImportPayload{}, shared.UsageError("--input is required")
-	}
-	formatValue, err := resolveMetadataKeywordImportFormat(inputValue, opts.Format)
-	if err != nil {
-		return keywordImportPayload{}, shared.UsageError(err.Error())
-	}
-
-	imported, err := readMetadataKeywordImportInput(inputValue, formatValue, strings.TrimSpace(opts.DefaultLocale))
-	if err != nil {
-		return keywordImportPayload{}, err
-	}
-
-	states, results, plans, issues, err := buildMetadataKeywordWriteResults(dirValue, versionValue, imported.locales, opts.Overwrite)
-	if err != nil {
-		return keywordImportPayload{}, err
-	}
-	if !opts.DryRun && len(issues) == 0 {
-		if err := ApplyWritePlans(plans); err != nil {
-			return keywordImportPayload{}, err
-		}
-	}
-	sideDataReportPath, sideDataRecordCount, err := maybeWriteMetadataKeywordSideDataReport(
-		dirValue,
-		versionValue,
-		inputValue,
-		formatValue,
-		opts.SideDataReportFile,
-		opts.DryRun || len(issues) > 0,
-		imported.sideData,
-	)
-	if err != nil {
-		return keywordImportPayload{}, err
-	}
-	detectedLocales := sortedKeys(imported.locales)
-
-	return keywordImportPayload{
-		states: states,
-		result: MetadataKeywordsImportResult{
-			Dir:                 dirValue,
-			Version:             versionValue,
-			Input:               inputValue,
-			Format:              formatValue,
-			DryRun:              opts.DryRun,
-			Valid:               len(issues) == 0,
-			DetectedLocales:     detectedLocales,
-			Results:             results,
-			Issues:              issues,
-			SideDataRecordCount: sideDataRecordCount,
-			SideDataReportPath:  sideDataReportPath,
-		},
-	}, nil
-}
-
-func maybeWriteMetadataKeywordSideDataReport(
-	dir string,
-	version string,
-	input string,
-	format string,
-	reportFile string,
-	dryRun bool,
-	records []MetadataKeywordSideDataRecord,
-) (string, int, error) {
-	if len(records) == 0 {
-		return "", 0, nil
-	}
-
-	path, err := resolveMetadataKeywordSideDataReportPath(dir, version, reportFile)
-	if err != nil {
-		return "", 0, err
-	}
-	if dryRun {
-		return path, len(records), nil
-	}
-
-	artifact := MetadataKeywordSideDataArtifact{
-		Dir:     dir,
-		Version: version,
-		Input:   input,
-		Format:  format,
-		Records: records,
-	}
-	data, err := encodeCanonicalJSON(artifact)
-	if err != nil {
-		return "", 0, err
-	}
-	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
-		return "", 0, err
-	}
-	if err := writeFileNoFollow(path, data); err != nil {
-		return "", 0, err
-	}
-	return path, len(records), nil
-}
-
-func resolveMetadataKeywordSideDataReportPath(dir string, version string, reportFile string) (string, error) {
-	if strings.TrimSpace(reportFile) != "" {
-		return strings.TrimSpace(reportFile), nil
-	}
-	resolvedVersion, err := validatePathSegment("version", version)
-	if err != nil {
-		return "", shared.UsageError(err.Error())
-	}
-	return filepath.Join(strings.TrimSpace(dir), "reports", "metadata-keywords-side-data", resolvedVersion+".json"), nil
-}
-
-func executeMetadataKeywordsLocalize(opts metadataKeywordsLocalizeOptions) (MetadataKeywordsLocalizeResult, error) {
-	dirValue, versionValue, err := validateMetadataKeywordDirVersion(opts.Dir, opts.Version)
-	if err != nil {
-		return MetadataKeywordsLocalizeResult{}, err
-	}
-	sourceLocale, err := validateMetadataKeywordLocale(opts.FromLocale)
-	if err != nil {
-		return MetadataKeywordsLocalizeResult{}, shared.UsageError(err.Error())
-	}
-	if len(opts.TargetLocales) == 0 {
-		return MetadataKeywordsLocalizeResult{}, shared.UsageError("--to-locales is required")
-	}
-
-	targets := make([]string, 0, len(opts.TargetLocales))
-	for _, rawLocale := range opts.TargetLocales {
-		resolvedLocale, localeErr := validateMetadataKeywordLocale(rawLocale)
-		if localeErr != nil {
-			return MetadataKeywordsLocalizeResult{}, shared.UsageError(localeErr.Error())
-		}
-		if strings.EqualFold(resolvedLocale, sourceLocale) {
-			return MetadataKeywordsLocalizeResult{}, shared.UsageError("--from-locale must be different from every target locale")
-		}
-		targets = append(targets, resolvedLocale)
-	}
-	sort.Strings(targets)
-
-	sourcePath, err := resolveExistingVersionLocalizationPath(dirValue, versionValue, sourceLocale)
-	if err != nil {
-		return MetadataKeywordsLocalizeResult{}, err
-	}
-	sourceLocalization, err := ReadVersionLocalizationFile(sourcePath)
-	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			return MetadataKeywordsLocalizeResult{}, shared.UsageErrorf("source locale file %s does not exist", sourcePath)
-		}
-		return MetadataKeywordsLocalizeResult{}, err
-	}
-	sourceKeywords, keywordErr := parseMetadataKeywordField(sourceLocalization.Keywords)
-	if keywordErr != nil {
-		return MetadataKeywordsLocalizeResult{}, shared.UsageErrorf("source locale %q has invalid keywords: %v", sourceLocale, keywordErr)
-	}
-
-	valuesByLocale := make(map[string][]string, len(targets))
-	for _, locale := range targets {
-		valuesByLocale[locale] = append([]string(nil), sourceKeywords...)
-	}
-
-	_, results, plans, issues, err := buildMetadataKeywordWriteResults(dirValue, versionValue, valuesByLocale, opts.Overwrite)
-	if err != nil {
-		return MetadataKeywordsLocalizeResult{}, err
-	}
-	if !opts.DryRun && len(issues) == 0 {
-		if err := ApplyWritePlans(plans); err != nil {
-			return MetadataKeywordsLocalizeResult{}, err
-		}
-	}
-
-	return MetadataKeywordsLocalizeResult{
-		Dir:             dirValue,
-		Version:         versionValue,
-		FromLocale:      sourceLocale,
-		DryRun:          opts.DryRun,
-		Valid:           len(issues) == 0,
-		DetectedLocales: targets,
-		Results:         results,
-		Issues:          issues,
-	}, nil
-}
-
-func executeMetadataKeywordsPlan(ctx context.Context, opts metadataKeywordsPlanOptions) (MetadataKeywordsPlanResult, error) {
-	resolvedAppID, versionValue, dirValue, platformValue, err := validateMetadataKeywordsRemoteInputs(opts.AppID, opts.Version, opts.Dir, opts.Platform)
-	if err != nil {
-		return MetadataKeywordsPlanResult{}, err
-	}
-	if opts.Apply && !opts.Confirm {
-		return MetadataKeywordsPlanResult{}, shared.UsageError("--confirm is required")
-	}
-
-	localState := opts.LocalState
-	if len(localState) == 0 {
-		localState, err = loadMetadataKeywordLocalState(dirValue, versionValue)
-		if err != nil {
-			return MetadataKeywordsPlanResult{}, err
-		}
-	}
-
-	client, err := shared.GetASCClient()
-	if err != nil {
-		return MetadataKeywordsPlanResult{}, fmt.Errorf("auth: %w", err)
-	}
-
-	requestCtx, cancel := shared.ContextWithTimeout(ctx)
-	defer cancel()
-
-	versionIDValue, _, err := resolveVersionID(requestCtx, client, resolvedAppID, versionValue, platformValue)
-	if err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			return MetadataKeywordsPlanResult{}, err
-		}
-		return MetadataKeywordsPlanResult{}, err
-	}
-
-	remoteVersionItems, err := fetchVersionLocalizations(ctx, client, versionIDValue)
-	if err != nil {
-		return MetadataKeywordsPlanResult{}, err
-	}
-
-	localPatches := keywordLocalStateToPatches(localState)
-	filteredRemoteItems := filterKeywordRemoteItems(remoteVersionItems, localState)
-	remoteVersion := remoteVersionItemsToVersionMap(filteredRemoteItems)
-	adds, updates, _, versionCalls := buildScopePlan(
-		versionDirName,
-		versionValue,
-		keywordPlanFields,
-		versionToPlanFields(localPatches),
-		versionToFieldMap(remoteVersion),
-	)
-	sortPlanItems(adds)
-	sortPlanItems(updates)
-	submitOpts := shared.SubmitReadinessOptions{}
-	if versionCreateWarningsNeedUpdateContext(localPatches, remoteVersion) {
-		readinessCtx, readinessCancel := shared.ContextWithTimeout(ctx)
-		submitOpts = shared.ResolveSubmitReadinessOptionsForVersionBestEffort(readinessCtx, client, versionIDValue, resolvedAppID, platformValue)
-		readinessCancel()
-	}
-	warnings := buildMetadataKeywordWarnings(localState, remoteVersion, submitOpts)
-
-	result := MetadataKeywordsPlanResult{
-		AppID:     resolvedAppID,
-		Version:   versionValue,
-		VersionID: versionIDValue,
-		Dir:       dirValue,
-		DryRun:    !opts.Apply || opts.DryRun,
-		Adds:      adds,
-		Updates:   updates,
-		APICalls:  buildAPICallSummary(scopeCallCounts{}, versionCalls),
-		Warnings:  warnings,
-	}
-
-	if !opts.Apply {
-		return result, nil
-	}
-
-	applySummary := applyMetadataKeywordChanges(ctx, client, versionIDValue, versionValue, localPatches, filteredRemoteItems)
-	result.DryRun = false
-	result.Total = applySummary.Total
-	result.Succeeded = applySummary.Succeeded
-	result.Failed = applySummary.Failed
-	result.Actions = applySummary.Actions
-	if applySummary.Failed == 0 {
-		result.Applied = true
-		return result, nil
-	}
-	result.Results = applySummary.Results
-	artifactPath, err := writeMetadataKeywordsApplyFailureArtifact(result, opts.FailureArtifactScope)
-	if err != nil {
-		return result, fmt.Errorf("write failure artifact: %w", err)
-	}
-	result.FailureArtifactPath = artifactPath
-	return result, nil
-}
-
-func applyMetadataKeywordChanges(
-	ctx context.Context,
-	client *asc.Client,
-	versionID string,
-	version string,
-	local map[string]versionLocalPatch,
-	remoteItems []asc.Resource[asc.AppStoreVersionLocalizationAttributes],
-) metadataKeywordsApplySummary {
-	remoteByLocale := make(map[string]remoteLocalizationState, len(remoteItems))
-	for _, item := range remoteItems {
-		locale := strings.TrimSpace(item.Attributes.Locale)
-		if locale == "" {
-			continue
-		}
-		remoteByLocale[locale] = remoteLocalizationState{
-			id: item.ID,
-			fields: versionFields(VersionLocalization{
-				Description:     item.Attributes.Description,
-				Keywords:        item.Attributes.Keywords,
-				MarketingURL:    item.Attributes.MarketingURL,
-				PromotionalText: item.Attributes.PromotionalText,
-				SupportURL:      item.Attributes.SupportURL,
-				WhatsNew:        item.Attributes.WhatsNew,
-			}),
-		}
-	}
-
-	locales := sortedLocaleUnion(local, remoteByLocale)
-	summary := metadataKeywordsApplySummary{
-		Actions: make([]ApplyAction, 0, len(locales)),
-		Results: make([]MetadataKeywordsMutationResult, 0, len(locales)),
-	}
-
-	for _, locale := range locales {
-		localPatch, localExists := local[locale]
-		remoteState, remoteExists := remoteByLocale[locale]
-		if !localExists {
-			continue
-		}
-
-		remoteFields := cloneStringMap(remoteState.fields)
-		adds, updates := countIntentChanges(keywordPlanFields, localPatch.setFields, remoteFields)
-		if adds == 0 && updates == 0 {
-			continue
-		}
-
-		result := MetadataKeywordsMutationResult{Locale: locale}
-		summary.Total++
-
-		switch {
-		case !remoteExists:
-			result.Action = "create"
-
-			createLoc := localPatch.localization
-			if hasVersionContent(localPatch.createLocalization) {
-				createLoc = localPatch.createLocalization
-			}
-
-			createCtx, createCancel := shared.ContextWithTimeout(ctx)
-			resp, createErr := client.CreateAppStoreVersionLocalization(createCtx, versionID, versionAttributes(locale, createLoc, true))
-			createCancel()
-			if createErr != nil {
-				result.Status = "failed"
-				result.Error = fmt.Sprintf(
-					"create version localization %s (fields: %s): %v",
-					locale,
-					formatAttemptedFieldMap(keywordPlanFields, localPatch.setFields),
-					createErr,
-				)
-				summary.Failed++
-				summary.Results = append(summary.Results, result)
-				continue
-			}
-
-			result.Status = "succeeded"
-			result.LocalizationID = strings.TrimSpace(resp.Data.ID)
-			summary.Succeeded++
-			summary.Actions = append(summary.Actions, ApplyAction{
-				Scope:          versionDirName,
-				Locale:         locale,
-				Version:        version,
-				Action:         "create",
-				LocalizationID: result.LocalizationID,
-			})
-		case remoteExists:
-			result.Action = "update"
-
-			updateCtx, updateCancel := shared.ContextWithTimeout(ctx)
-			resp, updateErr := client.UpdateAppStoreVersionLocalization(updateCtx, remoteState.id, versionAttributes(locale, localPatch.localization, false))
-			updateCancel()
-			if updateErr != nil {
-				result.Status = "failed"
-				result.LocalizationID = remoteState.id
-				result.Error = fmt.Sprintf(
-					"update version localization %s (fields: %s): %v",
-					locale,
-					formatAttemptedFieldMap(keywordPlanFields, localPatch.setFields),
-					updateErr,
-				)
-				summary.Failed++
-				summary.Results = append(summary.Results, result)
-				continue
-			}
-
-			result.Status = "succeeded"
-			result.LocalizationID = strings.TrimSpace(resp.Data.ID)
-			if result.LocalizationID == "" {
-				result.LocalizationID = remoteState.id
-			}
-			summary.Succeeded++
-			summary.Actions = append(summary.Actions, ApplyAction{
-				Scope:          versionDirName,
-				Locale:         locale,
-				Version:        version,
-				Action:         "update",
-				LocalizationID: result.LocalizationID,
-			})
-		}
-
-		summary.Results = append(summary.Results, result)
-	}
-
-	return summary
-}
-
-func writeMetadataKeywordsApplyFailureArtifact(result MetadataKeywordsPlanResult, scope string) (string, error) {
-	scope = strings.TrimSpace(scope)
-	if scope == "" {
-		scope = "metadata-keywords-apply"
-	}
-
-	artifact := metadataKeywordsApplyFailureArtifact{
-		AppID:       result.AppID,
-		Version:     result.Version,
-		VersionID:   result.VersionID,
-		Dir:         result.Dir,
-		Total:       result.Total,
-		Succeeded:   result.Succeeded,
-		Failed:      result.Failed,
-		GeneratedAt: time.Now().UTC().Format(time.RFC3339),
-		Results:     append([]MetadataKeywordsMutationResult(nil), result.Results...),
-	}
-
-	data, err := encodeCanonicalJSON(artifact)
-	if err != nil {
-		return "", err
-	}
-
-	path := filepath.Join(
-		".asc",
-		"reports",
-		scope,
-		fmt.Sprintf("failures-%d.json", time.Now().UTC().UnixNano()),
-	)
-	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
-		return "", err
-	}
-	if err := writeFileNoFollow(path, data); err != nil {
-		return "", err
-	}
-	return path, nil
-}
-
 func validateMetadataKeywordDirVersion(dir string, version string) (string, string, error) {
 	dirValue := strings.TrimSpace(dir)
 	if dirValue == "" {
@@ -1187,433 +121,6 @@ func validateMetadataKeywordLocale(locale string) (string, error) {
 		return "", fmt.Errorf("default locale is not supported for metadata keywords")
 	}
 	return resolvedLocale, nil
-}
-
-func resolveMetadataKeywordImportFormat(inputPath string, format string) (string, error) {
-	formatValue := strings.ToLower(strings.TrimSpace(format))
-	if formatValue == "" {
-		formatValue = keywordImportFormatAuto
-	}
-	switch formatValue {
-	case keywordImportFormatAuto:
-		if strings.TrimSpace(inputPath) == "-" {
-			return "", fmt.Errorf("--format is required when --input is -")
-		}
-		switch strings.ToLower(filepath.Ext(inputPath)) {
-		case ".csv":
-			return keywordImportFormatCSV, nil
-		case ".json":
-			return keywordImportFormatJSON, nil
-		case ".txt", ".text":
-			return keywordImportFormatText, nil
-		default:
-			return "", fmt.Errorf("could not infer input format from %q; use --format", inputPath)
-		}
-	default:
-		if _, ok := metadataKeywordImportFormats[formatValue]; ok {
-			return formatValue, nil
-		}
-		return "", fmt.Errorf("--format must be one of %s", strings.Join(metadataKeywordImportFormatList(), ", "))
-	}
-}
-
-func metadataKeywordImportFormatList() []string {
-	formats := []string{keywordImportFormatAuto}
-	for name := range metadataKeywordImportFormats {
-		formats = append(formats, name)
-	}
-	sort.Strings(formats[1:])
-	return formats
-}
-
-func readMetadataKeywordImportInput(inputPath, format, defaultLocale string) (metadataKeywordImportedData, error) {
-	data, err := readMetadataKeywordInputBytes(inputPath)
-	if err != nil {
-		return metadataKeywordImportedData{}, err
-	}
-	if len(strings.TrimSpace(string(data))) == 0 {
-		return metadataKeywordImportedData{}, shared.UsageError("import input is empty")
-	}
-
-	parser, ok := metadataKeywordImportFormats[format]
-	if !ok {
-		return metadataKeywordImportedData{}, shared.UsageErrorf("unsupported import format %q", format)
-	}
-	raw, err := parser(data, defaultLocale)
-	if err != nil {
-		return metadataKeywordImportedData{}, err
-	}
-	return normalizeImportedMetadataKeywords(raw, defaultLocale)
-}
-
-func readMetadataKeywordInputBytes(inputPath string) ([]byte, error) {
-	inputValue := strings.TrimSpace(inputPath)
-	if inputValue == "-" {
-		return io.ReadAll(os.Stdin)
-	}
-	return readFileNoFollow(inputValue)
-}
-
-func parseMetadataKeywordText(data []byte, defaultLocale string) (metadataKeywordImportedData, error) {
-	if strings.TrimSpace(defaultLocale) == "" {
-		return metadataKeywordImportedData{}, shared.UsageError("--locale is required for text input")
-	}
-	return metadataKeywordImportedData{
-		locales: map[string][]string{
-			defaultLocale: splitMetadataKeywordTokens(string(data)),
-		},
-	}, nil
-}
-
-func parseMetadataKeywordCSV(data []byte, defaultLocale string) (metadataKeywordImportedData, error) {
-	return parseMetadataKeywordCSVWithHeaders(
-		data,
-		defaultLocale,
-		[]string{"locale", "lang", "language"},
-		[]string{"keywords", "keywordfield", "keywordlist"},
-		[]string{"keyword", "term", "searchterm", "searchkeyword"},
-	)
-}
-
-func parseMetadataKeywordAstroCSV(data []byte, defaultLocale string) (metadataKeywordImportedData, error) {
-	return parseMetadataKeywordCSVWithHeaders(
-		data,
-		defaultLocale,
-		[]string{"locale", "lang", "language"},
-		nil,
-		[]string{"keyword"},
-	)
-}
-
-func parseMetadataKeywordCSVWithHeaders(
-	data []byte,
-	defaultLocale string,
-	localeHeaders []string,
-	keywordsHeaders []string,
-	keywordHeaders []string,
-) (metadataKeywordImportedData, error) {
-	reader := csv.NewReader(strings.NewReader(string(data)))
-	reader.FieldsPerRecord = -1
-	reader.TrimLeadingSpace = true
-
-	rows, err := reader.ReadAll()
-	if err != nil {
-		return metadataKeywordImportedData{}, shared.UsageErrorf("invalid csv input: %v", err)
-	}
-	if len(rows) == 0 {
-		return metadataKeywordImportedData{}, shared.UsageError("csv input is empty")
-	}
-
-	headerIndex := make(map[string]int, len(rows[0]))
-	for idx, header := range rows[0] {
-		headerIndex[normalizeMetadataKeywordHeader(header)] = idx
-	}
-
-	localeIdx, hasLocale := metadataKeywordHeaderIndex(headerIndex, localeHeaders...)
-	keywordsIdx, hasKeywords := metadataKeywordHeaderIndex(headerIndex, keywordsHeaders...)
-	keywordIdx, hasKeyword := metadataKeywordHeaderIndex(headerIndex, keywordHeaders...)
-	if !hasKeywords && !hasKeyword {
-		return metadataKeywordImportedData{}, shared.UsageError(`csv input requires a "keywords" or "keyword" column`)
-	}
-
-	result := make(map[string][]string)
-	sideData := make([]MetadataKeywordSideDataRecord, 0)
-	for rowIndex, row := range rows[1:] {
-		if metadataKeywordCSVRowEmpty(row) {
-			continue
-		}
-
-		rowLocale := strings.TrimSpace(defaultLocale)
-		if hasLocale && localeIdx < len(row) {
-			rowLocale = strings.TrimSpace(row[localeIdx])
-			if rowLocale == "" {
-				rowLocale = strings.TrimSpace(defaultLocale)
-			}
-		}
-		if rowLocale == "" {
-			return metadataKeywordImportedData{}, shared.UsageErrorf("csv row %d is missing a locale (set --locale or add a locale column)", rowIndex+2)
-		}
-
-		var tokens []string
-		if hasKeywords && keywordsIdx < len(row) {
-			tokens = append(tokens, splitMetadataKeywordTokens(row[keywordsIdx])...)
-		}
-		if len(tokens) == 0 && hasKeyword && keywordIdx < len(row) {
-			tokens = append(tokens, splitMetadataKeywordTokens(row[keywordIdx])...)
-		}
-		if len(tokens) == 0 {
-			continue
-		}
-		result[rowLocale] = append(result[rowLocale], tokens...)
-		fields := make(map[string]any)
-		for idx, rawHeader := range rows[0] {
-			if idx >= len(row) {
-				continue
-			}
-			value := strings.TrimSpace(row[idx])
-			if value == "" {
-				continue
-			}
-			if (hasLocale && idx == localeIdx) || (hasKeywords && idx == keywordsIdx) || (hasKeyword && idx == keywordIdx) {
-				continue
-			}
-			fields[normalizeMetadataKeywordHeader(rawHeader)] = value
-		}
-		if len(fields) > 0 {
-			sideData = append(sideData, MetadataKeywordSideDataRecord{
-				Locale:   rowLocale,
-				Keywords: append([]string(nil), tokens...),
-				Fields:   fields,
-			})
-		}
-	}
-	return metadataKeywordImportedData{
-		locales:  result,
-		sideData: sideData,
-	}, nil
-}
-
-func parseMetadataKeywordJSON(data []byte, defaultLocale string) (metadataKeywordImportedData, error) {
-	var payload any
-	if err := json.Unmarshal(data, &payload); err != nil {
-		return metadataKeywordImportedData{}, shared.UsageErrorf("invalid json input: %v", err)
-	}
-	result, err := collectMetadataKeywordJSON(payload, strings.TrimSpace(defaultLocale))
-	if err != nil {
-		return metadataKeywordImportedData{}, err
-	}
-	return result, nil
-}
-
-func collectMetadataKeywordJSON(payload any, defaultLocale string) (metadataKeywordImportedData, error) {
-	switch value := payload.(type) {
-	case []any:
-		result := make(map[string][]string)
-		sideData := make([]MetadataKeywordSideDataRecord, 0)
-		for idx, item := range value {
-			locale, keywords, fields, err := parseMetadataKeywordJSONObject(item, defaultLocale)
-			if err != nil {
-				return metadataKeywordImportedData{}, shared.UsageErrorf("json item %d: %v", idx, err)
-			}
-			result[locale] = append(result[locale], keywords...)
-			if len(fields) > 0 {
-				sideData = append(sideData, MetadataKeywordSideDataRecord{
-					Locale:   locale,
-					Keywords: append([]string(nil), keywords...),
-					Fields:   fields,
-				})
-			}
-		}
-		return metadataKeywordImportedData{locales: result, sideData: sideData}, nil
-	case map[string]any:
-		if nested, ok := value["localizations"]; ok {
-			return collectMetadataKeywordJSON(nested, defaultLocale)
-		}
-		if looksLikeMetadataKeywordLocalizationObject(value) {
-			locale, keywords, fields, err := parseMetadataKeywordJSONObject(value, defaultLocale)
-			if err != nil {
-				return metadataKeywordImportedData{}, err
-			}
-			result := metadataKeywordImportedData{
-				locales: map[string][]string{locale: keywords},
-			}
-			if len(fields) > 0 {
-				result.sideData = []MetadataKeywordSideDataRecord{{
-					Locale:   locale,
-					Keywords: append([]string(nil), keywords...),
-					Fields:   fields,
-				}}
-			}
-			return result, nil
-		}
-
-		result := make(map[string][]string)
-		sideData := make([]MetadataKeywordSideDataRecord, 0)
-		rawLocales := make([]string, 0, len(value))
-		for rawLocale := range value {
-			rawLocales = append(rawLocales, rawLocale)
-		}
-		sort.Strings(rawLocales)
-		for _, rawLocale := range rawLocales {
-			rawKeywords := value[rawLocale]
-			if object, ok := rawKeywords.(map[string]any); ok && looksLikeMetadataKeywordLocalizationObject(object) {
-				locale, keywords, fields, err := parseMetadataKeywordJSONObject(object, rawLocale)
-				if err != nil {
-					return metadataKeywordImportedData{}, shared.UsageErrorf("json locale %q: %v", rawLocale, err)
-				}
-				result[locale] = append(result[locale], keywords...)
-				if len(fields) > 0 {
-					sideData = append(sideData, MetadataKeywordSideDataRecord{
-						Locale:   locale,
-						Keywords: append([]string(nil), keywords...),
-						Fields:   fields,
-					})
-				}
-				continue
-			}
-			keywords, err := decodeMetadataKeywordValue(rawKeywords)
-			if err != nil {
-				return metadataKeywordImportedData{}, shared.UsageErrorf("json locale %q: %v", rawLocale, err)
-			}
-			result[rawLocale] = append(result[rawLocale], keywords...)
-		}
-		return metadataKeywordImportedData{locales: result, sideData: sideData}, nil
-	default:
-		return metadataKeywordImportedData{}, shared.UsageError("json input must be an object or array")
-	}
-}
-
-func looksLikeMetadataKeywordLocalizationObject(value map[string]any) bool {
-	if _, ok := value["locale"]; ok {
-		return true
-	}
-	for key := range value {
-		switch normalizeMetadataKeywordHeader(key) {
-		case "keywords", "keywordfield", "keywordlist", "keyword", "term", "searchterm", "searchkeyword":
-			return true
-		}
-	}
-	return false
-}
-
-func metadataKeywordJSONSideFields(value map[string]any) map[string]any {
-	fields := make(map[string]any)
-	for rawKey, rawValue := range value {
-		switch normalizeMetadataKeywordHeader(rawKey) {
-		case "locale", "keywords", "keywordfield", "keywordlist", "keyword", "term", "searchterm", "searchkeyword":
-			continue
-		default:
-			fields[rawKey] = rawValue
-		}
-	}
-	return fields
-}
-
-func parseMetadataKeywordJSONObject(raw any, defaultLocale string) (string, []string, map[string]any, error) {
-	object, ok := raw.(map[string]any)
-	if !ok {
-		return "", nil, nil, fmt.Errorf("expected object")
-	}
-
-	localeValue := strings.TrimSpace(defaultLocale)
-	if localeRaw, ok := object["locale"]; ok {
-		if localeString, ok := localeRaw.(string); ok {
-			localeValue = strings.TrimSpace(localeString)
-		} else {
-			return "", nil, nil, fmt.Errorf(`field "locale" must be a string`)
-		}
-	}
-	if localeValue == "" {
-		return "", nil, nil, fmt.Errorf("locale is required")
-	}
-	sideFields := metadataKeywordJSONSideFields(object)
-
-	for _, key := range []string{"keywords", "keywordField", "keywordList", "keyword", "term", "searchTerm", "searchKeyword"} {
-		if rawValue, ok := object[key]; ok {
-			keywords, err := decodeMetadataKeywordValue(rawValue)
-			if err != nil {
-				return "", nil, nil, err
-			}
-			return localeValue, keywords, sideFields, nil
-		}
-	}
-	for rawKey, rawValue := range object {
-		switch normalizeMetadataKeywordHeader(rawKey) {
-		case "keywords", "keywordfield", "keywordlist", "keyword", "term", "searchterm", "searchkeyword":
-			keywords, err := decodeMetadataKeywordValue(rawValue)
-			if err != nil {
-				return "", nil, nil, err
-			}
-			return localeValue, keywords, sideFields, nil
-		}
-	}
-	return "", nil, nil, fmt.Errorf("keywords field is required")
-}
-
-func decodeMetadataKeywordValue(raw any) ([]string, error) {
-	switch value := raw.(type) {
-	case string:
-		return splitMetadataKeywordTokens(value), nil
-	case []any:
-		tokens := make([]string, 0, len(value))
-		for idx, item := range value {
-			itemString, ok := item.(string)
-			if !ok {
-				return nil, fmt.Errorf("keyword item %d must be a string", idx)
-			}
-			tokens = append(tokens, splitMetadataKeywordTokens(itemString)...)
-		}
-		return tokens, nil
-	default:
-		return nil, fmt.Errorf("keywords must be a string or string array")
-	}
-}
-
-func normalizeImportedMetadataKeywords(raw metadataKeywordImportedData, defaultLocale string) (metadataKeywordImportedData, error) {
-	result := make(map[string][]string, len(raw.locales))
-	locales := make([]string, 0, len(raw.locales))
-	for locale := range raw.locales {
-		locales = append(locales, locale)
-	}
-	sort.Strings(locales)
-	for _, locale := range locales {
-		keywords := raw.locales[locale]
-		resolvedLocale := strings.TrimSpace(locale)
-		if resolvedLocale == "" {
-			resolvedLocale = strings.TrimSpace(defaultLocale)
-		}
-		if resolvedLocale == "" {
-			return metadataKeywordImportedData{}, shared.UsageError("locale is required for imported keywords")
-		}
-
-		canonicalLocale, err := validateMetadataKeywordLocale(resolvedLocale)
-		if err != nil {
-			return metadataKeywordImportedData{}, shared.UsageError(err.Error())
-		}
-
-		normalizedKeywords, err := normalizeMetadataKeywordTokensPreserveDuplicates(keywords)
-		if err != nil {
-			return metadataKeywordImportedData{}, shared.UsageErrorf("locale %q: %v", canonicalLocale, err)
-		}
-		result[canonicalLocale] = append(result[canonicalLocale], normalizedKeywords...)
-	}
-
-	for locale, keywords := range result {
-		normalizedKeywords, err := normalizeMetadataKeywordTokensPreserveDuplicates(keywords)
-		if err != nil {
-			return metadataKeywordImportedData{}, shared.UsageErrorf("locale %q: %v", locale, err)
-		}
-		result[locale] = normalizedKeywords
-	}
-	if len(result) == 0 {
-		return metadataKeywordImportedData{}, shared.UsageError("no keywords were found in the import input")
-	}
-	sideData := make([]MetadataKeywordSideDataRecord, 0, len(raw.sideData))
-	for _, record := range raw.sideData {
-		resolvedLocale := strings.TrimSpace(record.Locale)
-		if resolvedLocale == "" {
-			resolvedLocale = strings.TrimSpace(defaultLocale)
-		}
-		if resolvedLocale != "" {
-			canonicalLocale, err := validateMetadataKeywordLocale(resolvedLocale)
-			if err != nil {
-				return metadataKeywordImportedData{}, shared.UsageError(err.Error())
-			}
-			record.Locale = canonicalLocale
-		}
-		if len(record.Keywords) > 0 {
-			normalizedKeywords, err := normalizeMetadataKeywordTokensPreserveDuplicates(record.Keywords)
-			if err != nil {
-				return metadataKeywordImportedData{}, shared.UsageErrorf("locale %q: %v", record.Locale, err)
-			}
-			record.Keywords = normalizedKeywords
-		}
-		sideData = append(sideData, record)
-	}
-	return metadataKeywordImportedData{
-		locales:  result,
-		sideData: sideData,
-	}, nil
 }
 
 func buildMetadataKeywordWriteResults(dir, version string, valuesByLocale map[string][]string, overwrite bool) (map[string]keywordLocalState, []MetadataKeywordFileResult, []WritePlan, []MetadataKeywordIssue, error) {
@@ -1764,138 +271,6 @@ func resolveExistingVersionLocalizationPath(dir string, version string, canonica
 		return existingPath, nil
 	}
 	return canonicalPath, nil
-}
-
-func loadMetadataKeywordLocalState(dir, version string) (map[string]keywordLocalState, error) {
-	resolvedVersion, err := validatePathSegment("version", version)
-	if err != nil {
-		return nil, shared.UsageError(err.Error())
-	}
-	versionPath := filepath.Join(strings.TrimSpace(dir), versionDirName, resolvedVersion)
-	entries, err := os.ReadDir(versionPath)
-	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			return nil, shared.UsageErrorf("no version localization files found in %s", versionPath)
-		}
-		return nil, fmt.Errorf("failed to read %s: %w", versionPath, err)
-	}
-
-	states := make(map[string]keywordLocalState)
-	seenLocales := make(map[string]string)
-	for _, entry := range entries {
-		if entry.IsDir() || filepath.Ext(entry.Name()) != ".json" {
-			continue
-		}
-		rawLocale := strings.TrimSuffix(entry.Name(), ".json")
-		if strings.EqualFold(rawLocale, DefaultLocale) {
-			continue
-		}
-		locale, localeErr := validateMetadataKeywordLocale(rawLocale)
-		if localeErr != nil {
-			return nil, shared.UsageErrorf("invalid metadata keywords file %q: %v", entry.Name(), localeErr)
-		}
-		if err := recordCanonicalLocaleFile(seenLocales, locale, entry.Name()); err != nil {
-			return nil, shared.UsageError(err.Error())
-		}
-
-		path := filepath.Join(versionPath, entry.Name())
-		localization, readErr := ReadVersionLocalizationFile(path)
-		if readErr != nil {
-			return nil, shared.UsageErrorf("invalid metadata schema in %s: %v", path, readErr)
-		}
-		if strings.TrimSpace(localization.Keywords) == "" {
-			continue
-		}
-		states[locale] = buildMetadataKeywordLocalState(locale, path, localization)
-	}
-	if len(states) == 0 {
-		return nil, shared.UsageErrorf("no keyword metadata files with a non-empty keywords field were found in %s", versionPath)
-	}
-	return states, nil
-}
-
-func keywordLocalStateToPatches(states map[string]keywordLocalState) map[string]versionLocalPatch {
-	result := make(map[string]versionLocalPatch, len(states))
-	for locale, state := range states {
-		result[locale] = cloneVersionLocalPatch(state.patch)
-	}
-	return result
-}
-
-func filterKeywordRemoteItems(items []asc.Resource[asc.AppStoreVersionLocalizationAttributes], states map[string]keywordLocalState) []asc.Resource[asc.AppStoreVersionLocalizationAttributes] {
-	filtered := make([]asc.Resource[asc.AppStoreVersionLocalizationAttributes], 0, len(items))
-	for _, item := range items {
-		locale := strings.TrimSpace(item.Attributes.Locale)
-		if locale == "" {
-			continue
-		}
-		if _, ok := states[locale]; !ok {
-			continue
-		}
-		filtered = append(filtered, item)
-	}
-	return filtered
-}
-
-func remoteVersionItemsToVersionMap(items []asc.Resource[asc.AppStoreVersionLocalizationAttributes]) map[string]VersionLocalization {
-	result := make(map[string]VersionLocalization, len(items))
-	for _, item := range items {
-		locale := strings.TrimSpace(item.Attributes.Locale)
-		if locale == "" {
-			continue
-		}
-		result[locale] = NormalizeVersionLocalization(VersionLocalization{
-			Description:     item.Attributes.Description,
-			Keywords:        item.Attributes.Keywords,
-			MarketingURL:    item.Attributes.MarketingURL,
-			PromotionalText: item.Attributes.PromotionalText,
-			SupportURL:      item.Attributes.SupportURL,
-			WhatsNew:        item.Attributes.WhatsNew,
-		})
-	}
-	return result
-}
-
-func buildMetadataKeywordWarnings(states map[string]keywordLocalState, remote map[string]VersionLocalization, submitOpts shared.SubmitReadinessOptions) []MetadataKeywordsWarning {
-	patches := keywordLocalStateToPatches(states)
-	createWarnings := versionCreateWarningsForPatches(patches, remote, shared.SubmitReadinessCreateModePlanned, submitOpts)
-	warnings := make([]MetadataKeywordsWarning, 0, len(createWarnings))
-	for _, warning := range createWarnings {
-		warnings = append(warnings, MetadataKeywordsWarning{
-			Action:        "create",
-			Locale:        warning.Locale,
-			Message:       shared.FormatSubmitReadinessCreateWarning(warning),
-			MissingFields: append([]string(nil), warning.MissingFields...),
-		})
-	}
-	return warnings
-}
-
-func metadataKeywordHeaderIndex(headers map[string]int, names ...string) (int, bool) {
-	for _, name := range names {
-		if idx, ok := headers[name]; ok {
-			return idx, true
-		}
-	}
-	return 0, false
-}
-
-func normalizeMetadataKeywordHeader(value string) string {
-	normalized := strings.TrimSpace(strings.TrimPrefix(value, "\ufeff"))
-	normalized = strings.ToLower(normalized)
-	normalized = strings.ReplaceAll(normalized, " ", "")
-	normalized = strings.ReplaceAll(normalized, "-", "")
-	normalized = strings.ReplaceAll(normalized, "_", "")
-	return normalized
-}
-
-func metadataKeywordCSVRowEmpty(row []string) bool {
-	for _, item := range row {
-		if strings.TrimSpace(item) != "" {
-			return false
-		}
-	}
-	return true
 }
 
 func splitMetadataKeywordTokens(value string) []string {
@@ -2076,126 +451,6 @@ func printMetadataKeywordFileResultMarkdown(title string, results []MetadataKeyw
 	return nil
 }
 
-func printMetadataKeywordsPlanTable(result MetadataKeywordsPlanResult) error {
-	fmt.Printf("App ID: %s\n", result.AppID)
-	fmt.Printf("Version: %s\n", result.Version)
-	fmt.Printf("Dir: %s\n", result.Dir)
-	fmt.Printf("Dry Run: %t\n\n", result.DryRun)
-	if result.Applied {
-		fmt.Printf("Applied: %t\n\n", result.Applied)
-	}
-	if result.Total > 0 {
-		fmt.Printf("Succeeded: %d\n", result.Succeeded)
-		fmt.Printf("Failed: %d\n", result.Failed)
-		if result.FailureArtifactPath != "" {
-			fmt.Printf("Failure Artifact: %s\n", result.FailureArtifactPath)
-		}
-		fmt.Println()
-	}
-
-	pushResult := PushPlanResult{
-		AppID:    result.AppID,
-		Version:  result.Version,
-		Dir:      result.Dir,
-		DryRun:   result.DryRun,
-		Applied:  result.Applied,
-		Adds:     result.Adds,
-		Updates:  result.Updates,
-		APICalls: result.APICalls,
-		Actions:  result.Actions,
-	}
-	asc.RenderTable([]string{"change", "key", "scope", "locale", "version", "field", "reason", "from", "to"}, buildPlanRows(pushResult))
-	if len(result.APICalls) > 0 {
-		fmt.Println()
-		asc.RenderTable([]string{"operation", "scope", "count"}, buildAPICallRows(result.APICalls))
-	}
-	if len(result.Results) > 0 {
-		fmt.Println()
-		asc.RenderTable([]string{"locale", "action", "status", "localizationId", "error"}, buildMetadataKeywordMutationRows(result.Results))
-	} else if len(result.Actions) > 0 {
-		fmt.Println()
-		asc.RenderTable([]string{"scope", "locale", "version", "action", "localizationId"}, buildApplyActionRows(result.Actions))
-	}
-	if len(result.Warnings) > 0 {
-		fmt.Println()
-		asc.RenderTable([]string{"action", "locale", "message", "missingFields"}, buildMetadataKeywordWarningRows(result.Warnings))
-	}
-	return nil
-}
-
-func printMetadataKeywordsPlanMarkdown(result MetadataKeywordsPlanResult) error {
-	fmt.Printf("**App ID:** %s\n\n", result.AppID)
-	fmt.Printf("**Version:** %s\n\n", result.Version)
-	fmt.Printf("**Dir:** %s\n\n", result.Dir)
-	fmt.Printf("**Dry Run:** %t\n\n", result.DryRun)
-	if result.Applied {
-		fmt.Printf("**Applied:** %t\n\n", result.Applied)
-	}
-	if result.Total > 0 {
-		fmt.Printf("**Succeeded:** %d\n\n", result.Succeeded)
-		fmt.Printf("**Failed:** %d\n\n", result.Failed)
-		if result.FailureArtifactPath != "" {
-			fmt.Printf("**Failure Artifact:** %s\n\n", result.FailureArtifactPath)
-		}
-	}
-
-	pushResult := PushPlanResult{
-		AppID:    result.AppID,
-		Version:  result.Version,
-		Dir:      result.Dir,
-		DryRun:   result.DryRun,
-		Applied:  result.Applied,
-		Adds:     result.Adds,
-		Updates:  result.Updates,
-		APICalls: result.APICalls,
-		Actions:  result.Actions,
-	}
-	asc.RenderMarkdown([]string{"change", "key", "scope", "locale", "version", "field", "reason", "from", "to"}, buildPlanRows(pushResult))
-	if len(result.APICalls) > 0 {
-		fmt.Println()
-		asc.RenderMarkdown([]string{"operation", "scope", "count"}, buildAPICallRows(result.APICalls))
-	}
-	if len(result.Results) > 0 {
-		fmt.Println()
-		asc.RenderMarkdown([]string{"locale", "action", "status", "localizationId", "error"}, buildMetadataKeywordMutationRows(result.Results))
-	} else if len(result.Actions) > 0 {
-		fmt.Println()
-		asc.RenderMarkdown([]string{"scope", "locale", "version", "action", "localizationId"}, buildApplyActionRows(result.Actions))
-	}
-	if len(result.Warnings) > 0 {
-		fmt.Println()
-		asc.RenderMarkdown([]string{"action", "locale", "message", "missingFields"}, buildMetadataKeywordWarningRows(result.Warnings))
-	}
-	return nil
-}
-
-func buildMetadataKeywordWarningRows(warnings []MetadataKeywordsWarning) [][]string {
-	rows := make([][]string, 0, len(warnings))
-	for _, warning := range warnings {
-		rows = append(rows, []string{
-			warning.Action,
-			warning.Locale,
-			warning.Message,
-			strings.Join(warning.MissingFields, ","),
-		})
-	}
-	return rows
-}
-
-func buildMetadataKeywordMutationRows(results []MetadataKeywordsMutationResult) [][]string {
-	rows := make([][]string, 0, len(results))
-	for _, result := range results {
-		rows = append(rows, []string{
-			result.Locale,
-			result.Action,
-			result.Status,
-			result.LocalizationID,
-			result.Error,
-		})
-	}
-	return rows
-}
-
 func buildMetadataKeywordIssueRows(issues []MetadataKeywordIssue) [][]string {
 	rows := make([][]string, 0, len(issues))
 	for _, issue := range issues {
@@ -2217,26 +472,4 @@ func buildMetadataKeywordIssueRows(issues []MetadataKeywordIssue) [][]string {
 		})
 	}
 	return rows
-}
-
-func printMetadataKeywordsSyncTable(result MetadataKeywordsSyncResult) error {
-	if err := printMetadataKeywordFileResultTable("Keyword Import", result.Import.Results, result.Import.DetectedLocales, result.Import.Issues, result.Import.Dir, result.Import.Version, result.Import.DryRun, result.Import.SideDataRecordCount, result.Import.SideDataReportPath); err != nil {
-		return err
-	}
-	if result.Plan == nil {
-		return nil
-	}
-	fmt.Println()
-	return printMetadataKeywordsPlanTable(*result.Plan)
-}
-
-func printMetadataKeywordsSyncMarkdown(result MetadataKeywordsSyncResult) error {
-	if err := printMetadataKeywordFileResultMarkdown("Keyword Import", result.Import.Results, result.Import.DetectedLocales, result.Import.Issues, result.Import.Dir, result.Import.Version, result.Import.DryRun, result.Import.SideDataRecordCount, result.Import.SideDataReportPath); err != nil {
-		return err
-	}
-	if result.Plan == nil {
-		return nil
-	}
-	fmt.Println()
-	return printMetadataKeywordsPlanMarkdown(*result.Plan)
 }
